Move computer tool into .tool.beta namespace and rename internal image name.
