+++ docs/tools.qmd
@@ -438,12 +438,12 @@ The computer tool is based on the Anthropic [Computer Use](https://docs.anthropi
 
 ### Configuration
 
-The `computer()` tool runs within a Docker container. To use it with a task you need to reference the `inspect-computer-tool` image in your Docker compose file. For example:
+The `computer()` tool runs within a Docker container. To use it with a task you need to reference the `inspect-computer-tool-beta` image in your Docker compose file. For example:
 
 ``` {.yaml filename="compose.yaml"}
 services:
   default:
-    image: inspect-computer-tool
+    image: inspect-computer-tool-beta
 ```
 
 You can configure the container to not have Internet access as follows:
@@ -451,13 +451,13 @@ You can configure the container to not have Internet access as follows:
 ``` {.yaml filename="compose.yaml"}
 services:
   default:
-    image: inspect-computer-tool
+    image: inspect-computer-tool-beta
     network_mode: none
 ```
 
 Note that if you'd like to be able to view the model's interactions with the computer desktop in realtime, you will need to also do some port mapping to enable a VNC connection with the container. See the [VNC Client](#vnc-client) section below for details on how to do this.
 
-The `inspect-computer-tool` image is based on the [ubuntu:22.04](https://hub.docker.com/layers/library/ubuntu/22.04/images/sha256-965fbcae990b0467ed5657caceaec165018ef44a4d2d46c7cdea80a9dff0d1ea?context=explore) image and includes the following additional applications pre-installed:
+The `inspect-computer-tool-beta` image is based on the [ubuntu:22.04](https://hub.docker.com/layers/library/ubuntu/22.04/images/sha256-965fbcae990b0467ed5657caceaec165018ef44a4d2d46c7cdea80a9dff0d1ea?context=explore) image and includes the following additional applications pre-installed:
 
 - Firefox
 - VS Code
@@ -475,7 +475,7 @@ A task configured to use the computer tool might look like this:
 from inspect_ai import Task, task
 from inspect_ai.scorer import match
 from inspect_ai.solver import generate, use_tools
-from inspect_ai.tool import computer
+from inspect_ai.tool.beta import computer
 
 @task
 def computer_task():
@@ -511,7 +511,7 @@ You can use a [VNC](https://en.wikipedia.org/wiki/VNC) connection to the contain
 ``` {.yaml filename="compose.yaml"}
 services:
   default:
-    image: inspect-computer-tool
+    image: inspect-computer-tool-beta
     ports:
       - "5900"
       - "6080"
