+++ src/inspect_ai/tool/beta/__init__.py
@@ -1,0 +1,5 @@
+from ._computer import computer
+
+__all__ = [
+    "computer",
+]
