+++ src/inspect_ai/tool/beta/_computer/_common.py
@@ -123,12 +123,12 @@ async def computer_sandbox() -> SandboxEnvironment:
     else:
         raise PrerequisiteError(
             dedent("""
-                The computer tool service was not found in any of the sandboxes for this sample. Please add the computer tool service to your configuration. For example, the following Docker compose file uses the (currently internal) inspect-computer-tool image as its default sandbox:
+                The computer tool service was not found in any of the sandboxes for this sample. Please add the computer tool service to your configuration. For example, the following Docker compose file uses the (currently internal) inspect-computer-tool-beta image as its default sandbox:
 
                 services:
                   default:
                     # Temporary internal image until the official one is available
-                    image: "inspect-computer-tool"
+                    image: "inspect-computer-tool-beta"
                     init: true
                 """).strip()
         )
