+++ src/inspect_ai/tool/beta/_computer/_resources/Dockerfile
