+++ src/inspect_ai/tool/beta/_computer/_resources/computer_tool/_tool_result.py
