+++ src/inspect_ai/tool/beta/_computer/_resources/computer_tool/computer_tool.py
