+++ src/inspect_ai/tool/beta/_computer/_resources/image_home_dir/.config/tint2/tint2rc
