+++ src/inspect_ai/tool/beta/_computer/_resources/image_home_dir/novnc_startup.sh
