+++ src/inspect_ai/tool/beta/_computer/_resources/image_home_dir/xvfb_startup.sh
